# src/controllers/UserController.js
import UserService from '../services/userService.js';

class UserController {
    /**
     * @swagger
     * /users/{id}/update-v2:
     *   put:
     *     summary: Update a user's first and last name by ID (v2)
     *     tags:
     *       - Users
     *       - v2
     *     description: |
     *       Update only the first and last name of a user. This is a v2 endpoint.
     *     parameters:
     *       - in: path
     *         name: id
     *         required: true
     *         schema:
     *           type: integer
     *         description: The user ID
     *     requestBody:
     *       required: true
     *       content:
     *         application/json:
     *           schema:
     *             type: object
     *             properties:
     *               firstName:
     *                 type: string
     *               lastName:
     *                 type: string
     *     responses:
     *       200:
     *         description: User's name updated successfully
     *       404:
     *         description: User not found
     */
    updateUserNameV2 = (req, res) => {
        const userId = parseInt(req.params.id, 10);
        const { firstName, lastName } = req.body;
        if (!firstName || !lastName) {
            return res.status(400).json({ message: 'firstName and lastName are required' });
        }
        const updatedUser = this.userService.updateUserNameV2(userId, firstName, lastName);
        if (!updatedUser) {
            return res.status(404).json({ message: 'User not found' });
        }
        res.json(updatedUser);
    }
    userService;

    constructor() {
        this.userService = new UserService();
    }

    /**
     * @swagger
     * /users:
     *   post:
     *     summary: Create a new user
     *     tags:
     *       - Users
     *     requestBody:
     *       required: true
     *       content:
     *         application/json:
     *           schema:
     *             type: object
     *             properties:
     *               name:
     *                 type: string
     *               email:
     *                 type: string
     *     responses:
     *       201:
     *         description: User created successfully
     *       400:
     *         description: Invalid input
     */
    createUser = (req, res) => {
        const userData = req.body;
        const newUser = this.userService.createUser(userData);
        res.status(201).json(newUser);
    }

    /**
     * @swagger
     * /users/{id}:
     *   get:
     *     summary: Get a user by ID
     *     tags:
     *       - Users
     *     parameters:
     *       - in: path
     *         name: id
     *         required: true
     *         schema:
     *           type: integer
     *         description: The user ID
     *     responses:
     *       200:
     *         description: User found
     *       404:
     *         description: User not found
     */
    getUser = (req, res) => {
        const userId = parseInt(req.params.id, 10);
        const user = this.userService.getUser(userId);
        if (!user) {
            return res.status(404).json({ message: 'User not found' });
        }
        res.json(user);
    }

    /**
     * @swagger
     * /users/{id}:
     *   put:
     *     summary: Update a user by ID
     *     tags:
     *       - Users
     *       - Deprecated
     *     deprecated: true
     *     description: |
     *       This endpoint is deprecated and will be removed in future versions. 
     *       Please use the `/users/{id}/update` endpoint instead.
     *     parameters:
     *       - in: path
     *         name: id
     *         required: true
     *         schema:
     *           type: integer
     *         description: The user ID
     *     requestBody:
     *       required: true
     *       content:
     *         application/json:
     *           schema:
     *             type: object
     *             properties:
     *               name:
     *                 type: string
     *               email:
     *                 type: string
     *     responses:
     *       200:
     *         description: User updated successfully
     *       404:
     *         description: User not found
     */
    updateUser = (req, res) => {
        const userId = parseInt(req.params.id, 10);
        const updatedData = req.body;
        const updatedUser = this.userService.updateUser(userId, updatedData);

        if (!updatedUser) {
            return res.status(404).json({ message: 'User not found' });
        }

        // Add a custom header to indicate deprecation
        res.setHeader('X-Deprecation-Notice', 'This endpoint is deprecated and will be removed in future versions. Please use /users/{id}/update instead.');

        res.json(updatedUser);
    }

    /**
     * @swagger
     * /users/{id}/update:
     *   put:
     *     summary: Update a user by ID (v2)
     *     tags:
     *       - Users
     *       - v2
     *     description: |
     *       This is the updated version of the `PUT /users/{id}` endpoint.
     *       Use this endpoint for updating user details.
     *     parameters:
     *       - in: path
     *         name: id
     *         required: true
     *         schema:
     *           type: integer
     *         description: The user ID
     *     requestBody:
     *       required: true
     *       content:
     *         application/json:
     *           schema:
     *             type: object
     *             properties:
     *               name:
     *                 type: string
     *               email:
     *                 type: string
     *               phone:
     *                 type: string
     *                 description: The user's phone number
     *     responses:
     *       200:
     *         description: User updated successfully
     *       404:
     *         description: User not found
     */
    updateUserV2 = (req, res) => {
        const userId = parseInt(req.params.id, 10);
        const updatedData = req.body;
        const updatedUser = this.userService.updateUser(userId, updatedData);
        if (!updatedUser) {
            return res.status(404).json({ message: 'User not found' });
        }
        res.json(updatedUser);
    };

    /**
     * @swagger
     * /users/{id}:
     *   delete:
     *     summary: Delete a user by ID
     *     tags:
     *       - Users
     *     parameters:
     *       - in: path
     *         name: id
     *         required: true
     *         schema:
     *           type: integer
     *         description: The user ID
     *     responses:
     *       200:
     *         description: User deleted successfully
     *       404:
     *         description: User not found
     */
    deleteUser = (req, res) => {
        const userId = parseInt(req.params.id, 10);
        const deletedUser = this.userService.deleteUser(userId);
        if (!deletedUser) {
            return res.status(404).json({ message: 'User not found' });
        }
        res.json(deletedUser);
    }

    /**
     * @swagger
     * /users:
     *   get:
     *     summary: Get all users
     *     tags:
     *       - Users
     *     responses:
     *       200:
     *         description: A list of users
     *         content:
     *           application/json:
     *             schema:
     *               type: array
     *               items:
     *                 type: object
     *                 properties:
     *                   id:
     *                     type: integer
     *                   name:
     *                     type: string
     *                   email:
     *                     type: string
     */
    getAllUsers = (req, res) => {
        const users = this.userService.getAllUsers();
        res.json(users);
    }
}

export default UserController;